Updated create_payment profile from upstream
This change was necessary in order to properly create a payment profile.
The version we were using was outdated and did not factor in how Stripe
works.

A small modification was required in order to function properly:

`source.respond_to?(:imported)`

Our credit cards (in the test case) did not have this method.

Code was taken at around spree/spree:3e0b405a (upstream)
https://github.com/spree/spree/commit/3e0b405a#diff-d985c9327156568e9d480c22cfdac03eR188
but the method itself had gone over many changes prior to that so this
commit is not the only relevant one.