Fixed a bug that would leave an order permanently broken
Upon updating line_items the `OrderInventory` class moves around
inventory units. As a result, it will also delete shipments if they have
0 inventory units. Due to this, the order can be in payment or confirm
and have 0 shipments. Since the order is past delivery, shipments can
never be created again.

This happens because the method that is relied upon to reset the order's
state is `ensure_updated_shipments`, however since the act of updating
line items can delete all shipments, this has direct implications for
`ensure_updated_shipments` as it will only run on orders that have
shipments.

The fix is to grab the result from `ensure_updated_shipments`. If it is
`nil` then the `if` block prevented it from running. If it is `true`
then it's final executed method was `restart_checkout_flow` which is
essentially just calling `update_attributes` which itself returns a
boolean. Once we have the result we run `restart_checkout_flow`
"manually" if the result is `nil`. This will ensure that the order gets
kicked back to a state prior to delivery and new shipments will be
created upon traversing checkout.